requirements.txt
-i https://pypi.org/simple
anyio==4.6.2.post1; python_version >= '3.9'
beautifulsoup4==4.12.3; python_full_version >= '3.6.0'
certifi==2024.8.30; python_version >= '3.6'
charset-normalizer==3.4.0; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
fastcore==1.7.19; python_version >= '3.8'
fastlite==0.0.13; python_version >= '3.10'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.6; python_version >= '3.8'
httptools==0.6.4
httpx==0.27.2; python_version >= '3.8'
idna==3.10; python_version >= '3.6'
itsdangerous==2.2.0; python_version >= '3.8'
lucide-fasthtml==0.0.9; python_version >= '3.8'
lxml==5.3.0; python_version >= '3.6'
numpy==2.1.2; python_version >= '3.12'
oauthlib==3.2.2; python_version >= '3.6'
packaging==24.1; python_version >= '3.8'
pandas==2.2.3; python_version >= '3.9'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.1
python-fasthtml==0.8.0; python_version >= '3.10'
python-multipart==0.0.16; python_version >= '3.8'
pytz==2024.2
pyyaml==6.0.2
requests==2.32.3; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1; python_version >= '3.7'
soupsieve==2.6; python_version >= '3.8'
sqlite-minutils==3.37.0.post4; python_version >= '3.8'
starlette==0.41.2; python_version >= '3.8'
tzdata==2024.2; python_version >= '2'
urllib3==2.2.3; python_version >= '3.8'
uvicorn[standard]==0.32.0; python_version >= '3.8'
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
